[QUANTIZE] Refactor quantization codebase and fix model accuracy (#3543)
* Refactor.

* update

* update

* update

* update

* update

* update
